Server/requirements.txt
fastapi
pydantic
pillow
requests
torch
torchvision
open-clip-torch
scikit-learn
imageio
mysql-connector-python
uvicorn